+++ CHANGELOG.md
@@ -3,3 +3,16 @@
 * [2024-03-22]:
     - done: finished basic can bus
     - problems: shakehand not tested
+
+* [2024-03-22]:
+    - done: 
+        1. finished uart mode
+        2. add can bag handle
+    - tested:
+        1. opendevice
+        2. setCanRecvCallBack
+        3. enableCANMode
+        4. 
+    - problem:
+        1. the can baudrate setting reply is missing
+        2. heatbeats related do not work